+++ docs/src/pages/components/box.js
@@ -170,7 +170,7 @@ export default () => (
         <Link href="https://nulogy.design/tokens/">NDS theme</Link>
       </ListItem>
       <ListItem>
-        <Link href="https://storybook.nulogy.design/?selectedKind=Box">
+        <Link href="https://storybook.nulogy.design/?path=/story/box--box">
           View in Storybook
         </Link>
       </ListItem>