+++ docs/src/pages/components/card-set.js
@@ -50,7 +50,7 @@ export default () => (
       <SectionTitle>Resources</SectionTitle>
       <List>
         <ListItem>
-          <Link href="https://storybook.nulogy.design/?selectedKind=Cardset">
+          <Link href="https://storybook.nulogy.design/?path=/story/card--cardset">
             View in Storybook
           </Link>
         </ListItem>