+++ docs/src/pages/components/iconic-button.js
@@ -103,7 +103,7 @@ export default () => (
     <DocSection>
       <SectionTitle>Resources</SectionTitle>
       <ListItem>
-        <Link href="https://storybook.nulogy.design/?selectedKind=IconicButton">
+        <Link href="https://storybook.nulogy.design/?path=/story/iconicbutton--with-label">
           View in Storybook
         </Link>
       </ListItem>