+++ docs/src/pages/components/link.js
@@ -100,7 +100,7 @@ export default () => (
     <DocSection>
       <SectionTitle>Resources</SectionTitle>
       <ListItem>
-        <Link href="https://storybook.nulogy.design/?selectedKind=Link">
+        <Link href="https://storybook.nulogy.design/?path=/story/link--link">
           View in Storybook
         </Link>
       </ListItem>