+++ docs/src/pages/components/list.js
@@ -103,7 +103,7 @@ export default () => (
     <DocSection>
       <SectionTitle>Related components</SectionTitle>
       <ListItem>
-        <Link href="https://storybook.nulogy.design/?selectedKind=List">
+        <Link href="https://storybook.nulogy.design/?path=/story/list--list">
           View in Storybook
         </Link>
       </ListItem>